Enable infill pattern via infill line distance instead.
Fixes an issue where, when the global infill density was 0, and the settings for a mesh overlap changed the infill line distance (but not outright the density),  the infill pattern would be disabled in the mesh overlap settings. (You still will need to select both the density and line width to actually change the line width though.)

Relates to github issue Cura/10349 (not explicity linking it here, as it might close that issue then, while it shouldn't as most sub-issues in there are still unresolved one way or the other.