Always set the project name as the one provided by the user
There was an inconsistent behavior between saving a project file to disk (where the printer short name was readded after saving) and saving to the DigitalLibrary (where the printer short was NOT added again).

This commit fixes that by making sure that whatever the user put as the name of the file, will be the name of the project within Cura.

CURA-8566