Render material spool somewhat bigger and give it a margin
The image of the spool doesn't have a margin by itself, contrary to the other images which have a visual margin given by the size of the table. We have to give it a margin manually. This is one of the consequences of encoding margins inside of images that may be used all over the place. We can't just remove the margins inside of the image because the laptop image will be used in other places too.
Just adding a margin would subtract from the size of the spool image making it very small. To compensate I'm making the spool image slightly bigger.

Contributes to issue CURA-8609.