Allow word wrapping for error message
This prevents being able to align the text vertically with the troubleshooting button though. The text may become the larger of the two, in which case the text should be leading the height. Qt can do this with a Layout, however it then can't automatically adjust the width of the layout there since that would create a polish loop in the layout.

Contributes to issue CURA-8609.