Change distribution count to be the maximum diameter of distribution
Instead of the minimum. So previously a value of 2 would make it spread the error over 2 walls for an even number, or 3 for an odd. Now it makes it spread the error over 2 walls for an even number or 1 for an odd. This means that the minimum allowed value becomes 2, because for an even wall count it needs to place the error somewhere.

Done as a 5 minute fix.