Refactor: Use HttpRequestManager to async get token
This is a pretty complex structure with all the callbacks and stuff, both from the local web server and from the parsed reply.
What's worse, currently this gives an error due to threading conflicts. Apparently Firefox is making 10 requests on my computer. Probably because the reply is handled immediately without returning a proper response.

Contributes to issue CURA-8539.