(High) limits for number of walls to prevent engine crash.
The engine would reserve space for each outline that should be there according to the count, not for each wall that was actually printed. Values of 999999 where chosen after discussion for _some_ of these (but not anything to do wit hskirt-brim, since we use those in the frontend for dissalowed areas) becasue then you can just fill a number of nines to get a completely 'walled' model printed.