Don't attempt writing a workspace before there is a global stack
This is normally not possible from the interface. However on MacOS, the application menu gets carried to the top bar of the operating system. It is not blocked there, and the user could write a project file before they even add a printer. This prevents Cura from crashing when they do that.

Fixes Sentry issue CURA-2ZR.