Add button to visit plug-in website
There are a lot of buttons leading to websites now: An arrow leading to the author website. An author name leading to the author website. A 'read more' label leading to the plug-in website and this new button leading to the plug-in website. Maybe we should raise this with the designer.

Contributes to issue CURA-8565.