Show placeholder texts when compatibility lists are empty
When there are no compatible printers, we show that there is no information. After all, all materials should be compatible with some printer.
When there are no compatible support materials, we simply show 'None', because a material could be incompatible with all known support material types.

Contributes to issue CURA-8585.