Move What's New pages to cura-binary-data
These images are binary data, usually quite big and change often. We don't want to make the Git checkout unnecessarily huge with this. We'll put those properly in cura-binary-data.
The HTML files are not really binary, but they belong together so it's easy to administer them if they are in the same repository.