More expansive description for print head polygon setting
This description is shown for the X-min/Y-min/X-max/Y-max settings in the Machine Settings dialogue, in the tooltip. The new description is written to sort of apply to the polygon but also to those settings in the GUI. It needs to be understandable from the GUI in any case. Being technically correct for the polygon is bonus.

Fixes #11011.