Shorten translation for 'sync materials with printers'
These translations don't fit in the space we have for the button. We'll change the English text for the next release, but for this time we'll change just the translations that are giving us problems.

We'll accept that Czech and Turkish still have a bit of overlap in their buttons for this release.

Contributes to issue CURA-8753.