Get extruder values from appropriate sub-adhesion-extruder
The adhesion extruder is now split up into multiple sub-settings. Where applicable to the brim or skirt, get it from the skirt/brim extruder. Where applicable to a specific part of the raft, get it from that part's extruder. Where applicable to the whole raft, get it from the adhesion extruder.

Contributes to issue CURA-8868.