Increase canvas size of backplate image to stop repeating texture (MacOS does not attempt to tile the image, however Ubuntu does)
Texturing a set of faces with an image smaller than the faces will cause repeating. Normally this is not noticable since the edges of the backplate logos are mostly transparent but in this case the backplate faces were much taller than the input image. This change adds enough empty vertical height to the backplate image so that the tiling can not be seen.

CURA-7541