Fix clicking, and don't allow dragging if not needed
When I changed the MouseArea into a TooltipArea, the clicking broke because the tooltip area doesn't catch mouse events. I could add another mouse area on top, but for better efficiency it's nice to re-use the TooltipArea. Just need to remember to allow clicking it.
I also found a small issue where you could flick the table even if it's not big enough to scroll. This should fix that.

Contributes to issue CURA-8686.