Keep showing ip enter dialog after wrongly entered ip address
In component discover um 3 dialog. After a "complete" action the
dialog will immediately close. This was undesired behaviour as we want
to keep the dialog active if an IP address is wrongly added.

Additionally, it is no longer necessary to close/hide the dialog after
a reject/accept as this is done by default.

CURA-8955