Add "Ok" button in export/import materials popup dialog
Proposed solution takes a different approach compared to the previous
solution. Previously there was one message dialog that was hidden by
default. When the import/export material confirm dialog was triggered
the text and body of this message dialog would change.

I did deviate from this pattern as this might introduce hard to debug
issues when a property would be set on this dialog. As this dialog is
used for various instances every instance should update the same
properties. If one property is missed then the popup might show unwanted
information.

I propose a different solution instead. Here a popup is created
dynamically. This way the properties used for each instance is clear.
This might add a delay when showing the popup as the whole component
needs to be created when called.

CURA-8959