Don't make min single wall line width the same as the Min Odd Wall Line Width
The formula in that setting means that this setting was always the same as the Min Odd Wall Line Width,
but that is not really a natural value for this setting.
The absolute minimum for a line is generally known to be 75% of the nozzle size, so we use that instead.