reduce top bottom flow
The top bottom is a 100% density part of the print. For these parts it is super important to not over extrude. At the same time we dont want an underextruded top layer, since that is looking ugly. That is why we enable a roofing layer and reduce the flow for the rest of the top/bottom. PP-93