Credits given where credits due
The Update copyright feature of Pycharm automatically removed the
mentioning of one of our main contributors in this license header.
Added it manually again, unchecked the Update Copyright check box from
the commit window (on the second attempt) and order is restored to the
universe.

Mea Culpa @fieldOfView, Mea Maxima Culpa