reduce speed to match maximum flow in CURA413
With the increase in linewidth to the nozzle size, the material flow for especially the infill (but also inner walls) is increased. The speeds are reduced for the superdraft profiles to match the maximum material flow in CURA413. In the future we want to clean this up in a nice way, but for now this at least fixes some serious extrusion problems. PP-104