Scale convex hull only by X/Y shrinkage, and handle 0% gracefully
Only the X/Y shrinkage influences the convex hull collision area. Not the Z, nor the parent setting.

Also, don't scale the collision area to 0. This value is not allowed by the setting system, so it'd indicate an error when slicing, but before slicing the convex hull gets calculated which results in degenerate polygons. Don't do that. Instead, we'll just let it pretend the scale factor is 1.

Contributes to issue CURA-9091.