Fix the tooltip not re-sizing
I'm not entirely sure if this fixes it on windows, but it does at least fix things on Linux.

There are a few important changes to note here:
1. Increased the time of the fade timer to 200ms
2. Removed the mousearea, instead using the hover of scrollview

So why are these important? Nr 1 seems to be because updates seem to be faster? If I set it to 100
I could get it to work reasonably well on qt5, but in qt6, it's about 50% of the times that i'm too slow.
So we need the extra time.

As for the mouse area, it seems that the scrollview was eating up all the hover / containsMouse events.
As scrollview has the same properties that we need, we can just get rid of the mouseArea

CURA-9112