Make text area selection use theme colours
Also fix the highlight colour of the border. The highlight colour was changed in a later design but it was not changed here yet.
Without a theme colour it's going to use the system colours for the highlighting of text, which may be a very light grey that matches the background, making it impossible to see what part of the text is selected.

Hopefully fixes issue #12286.