We were previously both scaling down the buildplate and scaling up the disallowed areas of the models with shrinkage compensation. Doing both reduces the max usable area when scaling.
The solution is to do only one of these options.

Removing the disallowed area works fine when scaling models from their center. However we scale 2 or more models from the buildplate center rather than the model center. This means that the models can go outside the buildplate without it being obvious from the visualization.

The better solution is to remove the buildplate x/y scaling and keep the disallowed areas.

Removed shrinkage factor build volume scaling on x and y. This information is shown already by scaling the allowed areas around the objects.

Kept the bounding box scaling for z. This shows the max height a model can be. The objects disallowed area does not show this information.

Also removed scaling on non model disallowed areas on the build plate since this does not need to scale with the build plate anymore.

CURA-9271