Define Cura Python requirements
Our requirements.txt was never fully utilized.
For our builds we use a requirements.txt
specified in the cura-build-enviroment repo.
But that file contained all the Python requirements
including, testing and dev tools.

This requirements.txt should contain the bare minimum
for running Python. Since these dependencies are shipped
we need to specify the hashes as well.

These hashes aren't necessary for build and test requirements

Contributes to CURA-9365