Remove user and channel
Even though we had to patch these recipes, the user/channel
shouldn't be necesary any more because of the order in which
our remotes are defined in the conan-config repo. It will
first look into our own remotes, before going over to
conan-center-index.

Once the packages upstream are fixed for us, then we can use
the conan-center-index packages again, without changing the
code.

Contributes to CURA-9365