Catch wrong cloud responses when it responds with a new file name
It could happen if there are bugs in the cloud, e.g. if they strip special characters somewhere.
We'd want to know it if they do. But we also don't want Cura to crash if the remote connection is not reliable, or if there's changes/bugs in the future.

Fixes Sentry issue CURA-42F.