increase retract distance of PVA to default
PVA is stringing a lot due to the too small retract distance. This oozed material is causing under extrusion elsewhere, reducing the strength of the support. Too long retractions could cause reliability issues as well. However, since we reduced the number of retractions by ~40% for the smaller models (for bigger models it is ~0%, but then it is also not relevant since a lot of material is printed in between) with the new settings of PP-108 increasing the retract distance to 6.5mm is safe. PP-185