Store unmapped settings in a separate set
We don't want to create actual settings for all of these. Just store them in the profile metadata and restore them.
Serialising and deserialising these is a bit more work than what I could do now in 5 minutes so we'll have to do that a bit later, but at least for now we have the complete list.

Contributes to issue CURA-9432.