Fix path to icon file
The script got moved. It looks in the directory relative to __file__ for the icon and other files. Those other files also got moved, but the icons were placed in a centrally stored location. So the path needs to be updated to the new relative path.

Contributes to issue CURA-9409.