Set default for 3rd-party printers to Inside-Out
Overwhelmingly in our community it seems that the overhang and seam quality is better when printing inside-out.
Ultimaker found that the quality is better (in particular the dimensional accuracy) when printing outside-in. So let them have it outside-in and the rest inside-out.

Made after experimental evidence on Reddit (and in the past on Github).