Broaden scope of tags that get built
We want to build the beta tags too. Honestly, I think we could just build any tag automatically, but I'll be good and leave it at this.
I also broadened the tags to build for the stable releases to allow zeroes in the major version number, e.g. if we're going to release version 10.1.0.

Contributes to issue CURA-9365.