Build an AppImage with the latest Ubuntu
Instead of trying to support multiple libstdc++ or ship or own (bad pratcice in general)
an easy solution could be to provide two different AppImage's, were one is build on the
latest Ubuntu runner.

This should fix #11856 for the 5.1.0 release, but I don't consider this a permanent and/or
sustainable solution. We should still create a ticket on the backlog to solves this issue
in a more generic way and future-proof way.

Contributes to CURA-9205