Fix dark theme primary colour
The dark theme was overriding the primary colour with the wrong Ultimaker Blue colour. There is a newer blue, and it was applied as 'accent_1'. The dark theme had its own version of that one too. But the old Primary colour was still the wrong one.

Contributes to issue CURA-9465.