Explicitly build and copy OpenSSL shared
This is a workaround to ensure that we use and pack a shared library for OpenSSL 1.1.1l.
We currently compile OpenSSL statically for CPython, but our Python Dependenies (such as
PyQt6) require a shared library. Because Conan won't allow for building the same library
with two different options (easily) we need to install it explicitly and do a manual copy
to the VirtualEnv, such that Pyinstaller can find it.

Contributes to CURA-9365